+++ src/lib/overlays/drawers/MenuDrawer.tsx
@@ -13,7 +13,7 @@ import {
   IconList,
   IconSettings,
   IconStarFilled,
-  IconTools,
+  IconCalculator,
   IconTrendingUp,
   IconUpload,
   IconUser,
@@ -242,7 +242,7 @@ export function MenuDrawer({ collapsed }: { collapsed: boolean }) {
       items: [
         { key: AppPages.SHOWCASE, label: t('Tools.Showcase'), icon: <IconStarFilled size={16} /> },
         { key: AppPages.BENCHMARKS, label: t('Tools.Benchmarks'), icon: <IconLayoutGrid size={16} /> },
-        { key: AppPages.CALCULATORS, label: t('Tools.QuickUtils'), icon: <IconTools size={16} /> },
+        { key: AppPages.CALCULATORS, label: t('Tools.QuickUtils'), icon: <IconCalculator size={16} /> },
         { key: AppPages.WARP, label: t('Tools.WarpPlanner'), icon: <IconDiamond size={16} /> },
       ],
     },
