+++ src/lib/tabs/tabCalculators/AhaPanelContent.module.css
