+++ src/lib/tabs/tabCalculators/CalculatorPanel.module.css
