+++ src/lib/tabs/tabCalculators/CalculatorPanel.tsx
