+++ src/lib/tabs/tabCalculators/EhrPanelContent.module.css
