+++ src/lib/tabs/tabCalculators/calculatorPanels.ts
